Extract NavDropdown into reusable layout component
Make the hover dropdown generic with label, to, and children props
so it can be reused for other nav items with submenus.
